# src/doc/swagger/api.json
{"swagger":"2.0","info":{"description":"A simple SpringBoot application that provides API ","version":"1.0","title":"SpringBoot+SQLite application","contact":{"name":"RD offli","url":"http://localhost:2019"},"license":{"name":"Apache License Version 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:2019","basePath":"/","tags":[{"name":"monitor-controller","description":"Inspect monitoring data"},{"name":"team-controller","description":"teams operations"}],"paths":{"/football/teams":{"get":{"tags":["team-controller"],"summary":"List of available teams","operationId":"getTeamsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Team"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/football/teams/search":{"get":{"tags":["team-controller"],"summary":"Search team by title","operationId":"searchTeamUsingGET","produces":["application/json"],"parameters":[{"name":"teamTitle","in":"query","description":"teamTitle","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Team"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/football/teams/{id}":{"get":{"tags":["team-controller"],"summary":"Search team by ID","operationId":"getTeamByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Team"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/football/version":{"get":{"tags":["team-controller"],"summary":"Version","operationId":"versionUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/monitoring/data":{"get":{"tags":["monitor-controller"],"summary":"Print or download monitoring data","operationId":"getMonitoringDataUsingGET","produces":["application/json"],"parameters":[{"name":"format","in":"query","description":"format","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/MonitorData"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"MonitorData":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"memoryUsedHeap":{"type":"number","format":"double"},"memoryUsedNonHeap":{"type":"number","format":"double"},"processCpu":{"type":"number","format":"double"},"requestsPerMinute":{"type":"number","format":"double"},"systemCpu":{"type":"number","format":"double"},"threadsDaemon":{"type":"integer","format":"int32"},"threadsLive":{"type":"integer","format":"int32"},"timestamp":{"type":"string","format":"date-time"}},"title":"MonitorData"},"Team":{"type":"object","properties":{"club":{"type":"boolean"},"countryId":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"key":{"type":"string"},"national":{"type":"string"},"title":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"Team"}}}